Make script work with new version of CellMapper.
A big change in the website is that the code is now minified. Because of this, a simple find-and-replace no longer work to do many of the changes, so we now format the code of each function before searching for the places to change or insert code.
This uses the prettier and js-beautify libraries. I no longer remember why I used both, but they're used in various places. It would likely be simple to remove one of the libraries